Final Version of Wishlist. (AddToCart not fully working)
For some reason the add to cart functionality could not work. Using the same function that Raul created/used to add books from the book details, when I pass a book's id the server recognizes the number and passes it through the function and returns the cart's page. However, it does not add the book. An empty cart will remain empty from adding from the wishlist page. I have tried to figure this out but am struggling with the cart code.